backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
langchain==0.3.25
langchain-community==0.3.23
langchain-groq==0.3.2
langchain-huggingface==0.1.2
faiss-cpu==1.9.0.post1
sentence-transformers==3.0.1
pypdf==4.2.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1